Add redirect to hidden admin models
We don't want users to see the change page of profiles and specimens
directly, instead they should see it in the users and book change page,
respectively. Despite being hidden in the main navigation, there are
other ways to get to these unwanted pages (e.g. the related navigation
widget). This commit introduces redirects to assert the correct
visualization is shown.